# Implement `fold` on hash::table::{Iter,IterMut}.