@cristicbz 2016-10-23
Hmm, I was based on an old master, recompiling and running benchmarks now, I'll update.